cascade: split server-only loaders into *-loader files for client safety
